refactor: migrate matches, schedules, team goals
- Migrate 3 core business modules
- Organize by domain functionality
- Create proxies for all
